Move non-enemy code out of 'ai' directory
