+++ src/game_logic/entity_factory.cpp
@@ -22,7 +22,6 @@
 #include "engine/physics_system.hpp"
 #include "engine/sprite_tools.hpp"
 #include "game_logic/actor_tag.hpp"
-#include "game_logic/ai/blowing_fan.hpp"
 #include "game_logic/ai/blue_guard.hpp"
 #include "game_logic/ai/bomber_plane.hpp"
 #include "game_logic/ai/boss_episode_1.hpp"
@@ -32,21 +31,18 @@
 #include "game_logic/ai/hover_bot.hpp"
 #include "game_logic/ai/laser_turret.hpp"
 #include "game_logic/ai/messenger_drone.hpp"
-#include "game_logic/ai/missile.hpp"
 #include "game_logic/ai/prisoner.hpp"
 #include "game_logic/ai/red_bird.hpp"
 #include "game_logic/ai/rigelatin_soldier.hpp"
 #include "game_logic/ai/rocket_turret.hpp"
 #include "game_logic/ai/security_camera.hpp"
 #include "game_logic/ai/simple_walker.hpp"
-#include "game_logic/ai/sliding_door.hpp"
 #include "game_logic/ai/slime_blob.hpp"
 #include "game_logic/ai/slime_pipe.hpp"
 #include "game_logic/ai/smash_hammer.hpp"
 #include "game_logic/ai/snake.hpp"
 #include "game_logic/ai/spider.hpp"
 #include "game_logic/ai/spike_ball.hpp"
-#include "game_logic/ai/super_force_field.hpp"
 #include "game_logic/ai/watch_bot.hpp"
 #include "game_logic/behavior_controller.hpp"
 #include "game_logic/collectable_components.hpp"
@@ -54,13 +50,17 @@
 #include "game_logic/dynamic_geometry_components.hpp"
 #include "game_logic/effect_actor_components.hpp"
 #include "game_logic/effect_components.hpp"
-#include "game_logic/enemy_radar.hpp"
+#include "game_logic/interaction/blowing_fan.hpp"
 #include "game_logic/interaction/elevator.hpp"
+#include "game_logic/interaction/enemy_radar.hpp"
 #include "game_logic/interaction/force_field.hpp"
 #include "game_logic/interaction/locked_door.hpp"
+#include "game_logic/interaction/missile.hpp"
 #include "game_logic/interaction/respawn_checkpoint.hpp"
+#include "game_logic/interaction/sliding_door.hpp"
+#include "game_logic/interaction/super_force_field.hpp"
+#include "game_logic/interaction/tile_burner.hpp"
 #include "game_logic/item_container.hpp"
-#include "game_logic/tile_burner.hpp"
 #include "game_logic/trigger_components.hpp"
 
 #include <tuple>
