+++ src/game_logic/ingame_systems.hpp
@@ -30,7 +30,6 @@
 #include "game_logic/ai/prisoner.hpp"
 #include "game_logic/ai/rocket_turret.hpp"
 #include "game_logic/ai/simple_walker.hpp"
-#include "game_logic/ai/sliding_door.hpp"
 #include "game_logic/ai/slime_blob.hpp"
 #include "game_logic/ai/spider.hpp"
 #include "game_logic/ai/spike_ball.hpp"
@@ -40,8 +39,9 @@
 #include "game_logic/debugging_system.hpp"
 #include "game_logic/dynamic_geometry_system.hpp"
 #include "game_logic/effects_system.hpp"
-#include "game_logic/enemy_radar.hpp"
 #include "game_logic/interaction/elevator.hpp"
+#include "game_logic/interaction/enemy_radar.hpp"
+#include "game_logic/interaction/sliding_door.hpp"
 #include "game_logic/item_container.hpp"
 #include "game_logic/player.hpp"
 #include "game_logic/player/damage_system.hpp"
