+++ src/game_logic/interaction/sliding_door.cpp
