+++ src/game_logic/interaction/tile_burner.cpp
